Use exact-width integer types from stdint.h, no longer require long to be 32 bits.
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@7294 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/support/regression/tests/gcc-torture-execute-pr33779-1.c
@@ -8,9 +8,11 @@
 #pragma std_c99
 #endif
 
-int foo(long int i)
+#include <stdint.h>
+
+int foo(int32_t i)
 {
-  if (((unsigned long)(i + 1)) * 4 == 0)
+  if (((uint32_t)(i + 1)) * 4 == 0)
     return 1;
   return 0;
 }
